@@ -1,34 +1,41 @@
 plugins {
-    id 'org.springframework.boot' version '2.4.1'
-    id 'io.spring.dependency-management' version '1.0.10.RELEASE'
+    id 'org.springframework.boot' version '2.4.5'
+    id 'io.spring.dependency-management' version '1.0.11.RELEASE'
     id 'java'
 }
 
 group 'org.example'
 version '0.1'
-sourceCompatibility = '15'
+sourceCompatibility = '16'
 
 repositories {
     mavenCentral()
 }
 
 dependencies {
-    def springVersion = '2.4.1'
-    def lombokVersion = '1.18.16'
-    def jacksonVersion = '2.12.1'
+    def springBootVersion = '2.4.5'
+    def lombokVersion = '1.18.20'
+    def jacksonVersion = '2.12.3'
+    def postgresqlVersion = '42.2.20'
+    def junitVersion = '5.7.1'
 
     implementation 'org.springframework.boot:spring-boot-starter-web'
-    compile group: 'org.springframework.boot', name: 'spring-boot-starter-data-jpa', version: springVersion
-    compile group: 'org.springframework.boot', name: 'spring-boot-starter-thymeleaf', version: springVersion
+    implementation group: 'org.springframework.boot', name: 'spring-boot-gradle-plugin', version: springBootVersion
+    testImplementation group: 'org.springframework.boot', name: 'spring-boot-starter-test', version: springBootVersion
+    implementation group: 'org.springframework.boot', name: 'spring-boot-starter-data-jpa', version: springBootVersion
+    implementation group: 'org.springframework.boot', name: 'spring-boot-starter-thymeleaf', version: springBootVersion
 
     compileOnly group: 'org.projectlombok', name: 'lombok', version: lombokVersion
     annotationProcessor group: 'org.projectlombok', name: 'lombok', version: lombokVersion
 
-    compile group: 'com.fasterxml.jackson.core', name: 'jackson-core', version: jacksonVersion
-    compile group: 'com.fasterxml.jackson.core', name: 'jackson-annotations', version: jacksonVersion
-    compile group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: jacksonVersion
+    implementation group: 'com.fasterxml.jackson.core', name: 'jackson-core', version: jacksonVersion
+    implementation group: 'com.fasterxml.jackson.core', name: 'jackson-annotations', version: jacksonVersion
+    implementation group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: jacksonVersion
 
-    compile group: 'org.postgresql', name: 'postgresql', version: '42.2.18'
+    implementation group: 'org.postgresql', name: 'postgresql', version: postgresqlVersion
+
+    testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: junitVersion
+    testRuntimeOnly group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: junitVersion
 
     testImplementation('org.springframework.boot:spring-boot-starter-test') {
         exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
